avocado.virt: Figure out guest image path if none provided
It is a common use case that people just want to use a particular
guest image file on a test. That can be done using the newly
introduced test runner option --guest-image-path. This value
is going to be passed to the test, and in case none is provided,
use what the virt.defaults module tells we should use.

With this we get extra flexibility and convenience running our
tests.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>