avocado-virt: Introduce a namespace for the plugin variables
Instead of setting flat, global variables for the options
passed to the avocado command line, introduce a namespace -
avocado.args.run. Examples:

    avocado.args.run.qemu_bin
    avocado.args.run.qemu_dst_bin
    avocado.args.run.guest_user

This will reduce the chance of variable clashes inside
plugin code to virtually zero. Change all references
to those variables, and add the missing qemu_bin and
qemu_io command line options to the command line, so
we make all options to have an equivalent in the command
line.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>