avocado.virt.qemu.machine: Add screendump thread to VMs
Create a screendump thread on every VM (optional,
customizable through avocado.args.run.screendump_thread.enable).
This is a departure from previous efforts (virt-test), where
the thread was not under class control.

So at the QEMU process init stage, start the thread, when
the VM is shut down, stop it. This way, we have a visual record
of what was going on in the VM during the test process.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>