API review: sync with the avocado API changes
While modules living under avocado.core are not intended to be
used by tests, at this point extensions are OK using it. We may,
in the (far?) future, come with an extension/plugin API definition
but this is currently out of scope.

So, for now, let's sync avocado-virt with the proposed changes on
avocado itself.

These are the changes that impact avocado-virt:

* rename of avocado.test.Test to avocado.Test
* rename of avocado.aexpect to avocado.core.aexpect
* rename of avocado.data_dir to avocado.core.data_dir
* rename of avocado.settings to avocado.core.settings

Signed-off-by: Cleber Rosa <crosa@redhat.com>