avocado.virt: Set defaults in param tree instead of test params
The plugin defaults like "qemu_bin" used to be set on Test.__init__.
This is impossible with the new read-only avocado params. What this
patch does instead is it injects those defaults into the params in
"/run" (in order to be visible to "params.get(path='*')").

The benefit of this is it only happens once per all jobs.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>