SPEC file: require avocado package of the same version
So far the RPM packages have not been strict about the version
of avocado that they require.  The same version has probably
been used because of newer package availability on the repos.

Anyway, let's require the same package versions so that users
may not be surprised with breakage because of different version
and interfaces.

Signed-off-by: Cleber Rosa <crosa@redhat.com>