qemu: Use KVM if possible
On Fedora 24, the boot test hangs for me. Adding -enable-kvm to the QEMU
command line fixes it. While I haven't digged into why it hangs with
TCG, adding -enable-kvm if possible is useful nevertheless.

Test the accessibility of /dev/kvm file, which is the universal entry to
KVM, and add -enable-kvm to the command line if available.

Also add a line of log once the decision is done.

Thanks to Amador Pahim <apahim@redhat.com> for the help on this patch.

Signed-off-by: Fam Zheng <famz@redhat.com>